requirements.txt
pytesseract
easyocr
pypdf
pdfminer.six
python-docx
google-generativeai
pandas
openpyxl
openai
PyMuPDF